+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+##
+### Fixed
+ - Access query fixed to work with elemMatch.
+
 ## 0.6.6
 ### Fixed
  - Access role tests causing crashes.